jq-rally-api/requirements.txt
asgiref==3.3.4
azure-common==1.1.27
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
certifi==2020.4.5.1
cffi==1.14.6
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2
django-appconf==1.0.4
django-cors-headers==3.7.0
django-filter==2.4.0
django-imagekit==4.0.2
django-rest-swagger==2.2.0
django-storages==1.11.1
django-toolbelt==0.0.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
drf-writable-nested==0.6.2
gunicorn==20.1.0
idna==2.9
itypes==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==1.4.6
openapi-codec==1.3.2
pilkit==2.0
Pillow==8.3.1
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2019.3
requests==2.23.0
simplejson==3.17.0
six==1.16.0
sqlparse==0.3.1
static3==0.7.0
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.25.9
whitenoise==5.2.0
